Update Final Ass2.m
I've made the outputBaseFolder a global variable so the output from the pre-processing will have some to go. It was limited to only cropping before.
I've added my code for the salt & pepper and gaussian noise.
I've changed the code for histogram equalization so it will work with the outputBaseFolder.